+++ ambra_sdk/__init__.py
@@ -1,8 +1,8 @@
-__version__ = '3.20.5.0-1'
+__version__ = '3.20.6.0-1'
 
 # from footer in api.html
-API_VERSION = 'LBL0022 v39.0 2020-07-15'
+API_VERSION = 'LBL0022 v40.0 2020-08-26'
 # equal for API_VERSION based on dictionary.html
-MODELS_VERSION = 'LBL0022 v39.0 2020-07-15'
+MODELS_VERSION = 'LBL0022 v40.0 2020-08-26'
 # from storage_api.html
-STORAGE_VERSION = 'LBL0038 v9.0 2020-06-03'
+STORAGE_VERSION = 'LBL0038 v10.0 2020-08-26'