+++ tests/test_ambra.py
@@ -2,12 +2,12 @@ from ambra_sdk import API_VERSION, MODELS_VERSION, STORAGE_VERSION
 
 
 def test_api_version():
-    assert API_VERSION == 'LBL0022 v39.0 2020-07-15'
+    assert API_VERSION == 'LBL0022 v40.0 2020-08-26'
 
 
 def test_models_version():
-    assert MODELS_VERSION == 'LBL0022 v39.0 2020-07-15'
+    assert MODELS_VERSION == 'LBL0022 v40.0 2020-08-26'
 
 
 def test_storage_version():
-    assert STORAGE_VERSION == 'LBL0038 v9.0 2020-06-03'
+    assert STORAGE_VERSION == 'LBL0038 v10.0 2020-08-26'